pyproject.toml
[tool.poetry]
name = "pymc-mlflow"
version = "0.1.0"
description = ""
authors = ["James Bristow <jbris@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
mlflow = "^2.11.1"
pymc = "^5.10.4"
arviz = "^0.17.0"
scikit-learn = "^1.4.1.post1"
numpy = "^1.26.4"
pandas = "^2.2.1"
scipy = "^1.12.0"
matplotlib = "^3.8.3"
boto3 = "^1.34.59"
botocore = "^1.34.59"
psycopg2 = "^2.9.9"
plotly = "^5.19.0"
prometheus-flask-exporter = "^0.23.0"
bentoml = "^1.2.6"
hydra-core = "^1.3.2"
pydantic = "^2.6.3"
pymc-experimental = {git = "https://github.com/pymc-devs/pymc-experimental.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
